change: Quiz 관련 interface를 d.ts로 분리
